LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/merge-span-with-style-after-forwarddelete-having-contenteditable-expected.txt

Back

space


FAIL waiting for command to execute assert_equals: Style is not preserved for the span after pressing backspace in contenteditable expected "<h1><span style=\"background-color:red;\" id=\"back\">Back</span><span style=\"background-color: red;\">space</span></h1>" but got "<h1><span style=\"background-color:red;\" id=\"back\">Back</span></h1>\n<h1><span style=\"background-color: red;\">space</span></h1>"
